Multiple fixes re: previous commit.
- Add dataset to __init__ file
- Fix loss function for sem seg to instance-weighted x-ent from class-weighted x-ent
- Remove point cloud augmentation from yacs files (creates conflicts with sample weights during training)
- Add torch.long() casts to loss and metric fxns for sem seg for label tensors
- Add dimension transpose in ScanNet dataset for proper dimension ordering

+++ shaper/models/metric.py
@@ -51,7 +51,7 @@ class SegAccuracy(nn.Module):
 
     def forward(self, preds, labels):
         seg_logit = preds["seg_logit"]
-        seg_label = labels["seg_label"]
+        seg_label = labels["seg_label"].long()
         pred_label = seg_logit.argmax(1)
 
         # (batch_size, num_points)
